requirements.txt
pandas
sqlalchemy
mysql-connector-python
langchain
langchain-community
langchain-openai
streamlit
plotly